refactor(board): extract pure layout geometry into board/boardLayout.ts (D2 Tier-0)
`Board.tsx` (1266 LOC) carried ~470 lines of React-free geometry between
its imports and the component: `computeBoardLayout` / `layoutBranches` /
`calculateBoardFitScale` + the `BoardLayout` / `BoardLayoutTile` / `LayoutZone`
/ `HubLookup` types + the `TILE_UNIT` constants. Moved verbatim to
`client/src/components/board/boardLayout.ts` (483 LOC, under the max-lines
budget); `Board.tsx` imports the two functions + the `BoardLayout` type it
uses and drops to 797 LOC.

Zero behaviour change — a pure move, proven by `tsc -b` and the existing
`Board.reviewContainment.test.tsx` (repointed to the new module) +
`boardDiagnostics.test.ts`. The geometry is now independently testable
without mounting a component.

This is D2's free Tier-0 only (`REFACTOR_OPPORTUNITIES.md` §3 D2 / scoping
doc). The component hook-decomposition (useBoardCamera etc.) is NOT touched
— it needs board-interaction e2e coverage first.
